# stat/distuv: improve Poisson.Rand performance for Lambda >= 10