--- a/stat/distuv/poisson_test.go
+++ b/stat/distuv/poisson_test.go
@@ -121,7 +121,7 @@ func TestPoisson(t *testing.T) {
 func testPoisson(t *testing.T, p Poisson, i int) {
 	const (
 		tol = 1e-2
-		n   = 1e6
+		n   = 2e6
 	)
 	x := make([]float64, n)
 	generateSamples(x, p)