Please take a look.

benchstat:
```
name                old time/op  new time/op  delta
PoissonRand/case_0   297ns ± 5%    76ns ±10%  -74.56%  (p=0.000 n=10+10)
PoissonRand/case_1   284ns ± 6%   107ns ± 7%  -62.21%  (p=0.000 n=10+10)
PoissonRand/case_2   284ns ± 6%   118ns ± 4%  -58.34%  (p=0.000 n=10+10)
PoissonRand/case_3   236ns ± 3%   239ns ± 6%     ~     (p=0.381 n=10+10)
PoissonRand/case_4   105ns ± 6%   107ns ± 8%     ~     (p=0.236 n=10+10)
PoissonRand/case_5  73.8ns ± 4%  75.1ns ± 4%     ~     (p=0.148 n=10+10)
PoissonRand/case_6  60.3ns ± 5%  60.9ns ± 5%     ~     (p=0.307 n=10+10)
```

<!--
Checklist:

- API changes have been discussed
- code is goformated correctly (goimports)
- packages with generated code have had code generation run
- tests pass locally
- linked to relevant issues

Please make sure your commit message summary line and pull request
title match the Go convention; a one-line summary of the change,
prefixed by the primary affected package that should complete the
sentence, "This change modifies Gonum to _____."
-->